[#16] 라우트 관련 파일 정리
- AppContent 컴포넌트 제거
- routes.ts 파일 제거
- 라우트 관련 로직 App.tsx로 전부 이동
